reorganize database extensions
