+++ AppBookDataAnalyzer/Database/Question/Database+MultipleChoice.swift
@@ -30,4 +30,13 @@ extension Database {
         }
     }
     
+    struct MultipleChoiceMultipleAnswersAllowed {
+        static let tableName = "mc_multiple_answers_allowed"
+        struct Column {
+            static let id = "id"
+            static let questionId = "quiz_id"
+            static let maxSelectedAnswersAllowed = "max_selected_answers_allowed"
+        }
+    }
+    
 }
