+++ AppBookDataAnalyzer/Database/Question/Database+Question.swift
@@ -25,4 +25,14 @@ extension Database {
         }
     }
     
+    struct QuestionPageJoin {
+        static let tableName = "question_to_page_join"
+        struct Column {
+            static let id = "id"
+            static let questionId = "question_id"
+            static let appbookId = "appbook_id"
+            static let pageNumber = "page_number"
+        }
+    }
+    
 }
